requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
pydantic==2.8.2
httpx==0.27.0
python-dotenv==0.21.0
pyyaml==6.0.1
mss==9.0.1
pytesseract==0.3.10
Pillow==10.4.0
pywinauto==0.6.8
numpy==1.26.4
opencv-python==4.10.0.84
speechrecognition==3.10.0
PyAudio==0.2.14
scipy==1.13.1
googlesearch-python==1.2.5
beautifulsoup4==4.12.3
requests==2.32.3
youtube-transcript-api==0.6.2
pdfplumber==0.11.4
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
spacy==3.7.2
selenium==4.23.1
celery==5.4.0
langchain==0.2.10
sentence-transformers==3.0.1